Добавила модуляцию и импорты-экспорты
